converted test algebraic expression to use GoogleTest
